feat: auth protect

+++ src/app/api/backlogs/[id]/items/[itemId]/route.ts
@@ -1,27 +1,35 @@
 import { deleteBacklogItem, putBacklogItem } from "@/services/backlogItem";
+import { isAuthorizedBacklogOwner } from "@/services/backlogs";
 import { sendMsg } from "@/utils";
 import { revalidateTag } from "next/cache";
-import { NextRequest, NextResponse } from "next/server";
+import { NextRequest } from "next/server";
 
 export async function DELETE(
   request: NextRequest,
   { params: { id, itemId } }: { params: { id: string; itemId: string } },
 ) {
   try {
+    const isAuthorize = await isAuthorizedBacklogOwner(id);
+    if (isAuthorize) return sendMsg.error("Not authorized", 401);
     await deleteBacklogItem(itemId);
     revalidateTag(`backloglist${id}`);
-    return NextResponse.json({ message: "Deleted" }, { status: 202 });
+    return sendMsg.success("Deleted", 202);
   } catch (error) {
     sendMsg.error(error);
   }
 }
 
-export async function PUT(request: NextRequest) {
+export async function PUT(
+  request: NextRequest,
+  { params: { id } }: { params: { id: string } },
+) {
   const data = await request.json();
   try {
+    const isAuthorize = await isAuthorizedBacklogOwner(id);
+    if (isAuthorize) return sendMsg.error("Not authorized", 401);
     await putBacklogItem(data);
-    return NextResponse.json("Success");
+    return sendMsg.success();
   } catch (error) {
-    throw new Error(`${error}`);
+    sendMsg.error(error);
   }
 }
